Voyage calculator refactor
Adds ability to toggle different layouts for recommended lineups (TODO: save user preference across sessions). Stylesheets have been updated accordingly.

Moves recommended note to the top of lineup tab. A larger rewrite of the note is forthcoming.

Moves current voyage toggle button to the bottom of the tool. Additional changes to better display messages about ongoing voyages is forthcoming.

Refactors voyage calculator into smaller subcomponents, relocates voyagehelpers and voyagestats, for easier maintenance. Removes crewpopup.tsx, which has been replaced in lineupviewer.

Other things planned:
* Ability to dismiss recommended lineup
* Ability to re-estimate recommended lineup (maybe)
* Remember requests and results in session (maybe); this would address the complaint that recommendations are lost when resizing browser window
