+++ src/components/voyagecalculator.tsx
@@ -1,19 +1,16 @@
 import React from 'react';
-import { isMobile } from 'react-device-detect';
-import { Icon, Modal, Form, Button, Dropdown, Table, Message, Checkbox, Select, Segment, Header, Image, Tab, Grid, Card, Popup } from 'semantic-ui-react';
-import { Link } from 'gatsby';
+import { Icon, Button, Message, Image, Grid, Card } from 'semantic-ui-react';
 
 import CONFIG from './CONFIG';
 import allTraits from '../../static/structured/translation_en.json';
 
-import { VoyageStats } from '../components/voyagestats';
+import ConfigEditor from '../components/voyagecalculator/configeditor';
+import Recommender from '../components/voyagecalculator/recommender';
+import CIVASMessage from '../components/voyagecalculator/civas';
+import { VoyageStats } from '../components/voyagecalculator/voyagestats';
 
-import { guessCurrentEvent, getEventData } from '../utils/events';
 import { mergeShips } from '../utils/shiputils';
 import { useStateWithStorage } from '../utils/storage';
-import { CALCULATORS, CalculatorState } from '../utils/voyagehelpers';
-
-import UnifiedWorker from 'worker-loader!../workers/unifiedWorker';
 
 const AllDataContext = React.createContext();
 
@@ -110,9 +107,9 @@ const VoyageCalculator = (props: VoyageCalculatorProps) => {
 			if (myShip) {
 				myShip.id = ship.id;	// VoyageStats needs ship id to identify ship on existing voyage
 				myShip.index = { left: (ownedCount-idx+1), right: idx-1 };
-				if (idx == 0)
+				if (idx === 0)
 					myShip.index = { left: 1, right: ownedCount-1 };
-				else if (idx == 1)
+				else if (idx === 1)
 					myShip.index = { left: 0, right: 0 };
 			}
 		});
@@ -152,40 +149,41 @@ const VoyageMain = (props: VoyageMainProps) => {
 
 	return (
 		<React.Fragment>
-			{voyageState == 'input' &&
-				<Grid columns={2} stackable>
-					<Grid.Column width={14}>
-						<Card.Group>
-							<Card>
-								<Card.Content>
-									<Image floated='right' src={`${process.env.GATSBY_ASSETS_URL}atlas/icon_${voyageConfig.skills.primary_skill}.png`} style={{ height: '2em' }} />
-									<Card.Header>{CONFIG.SKILLS[voyageConfig.skills.primary_skill]}</Card.Header>
-									<p>primary</p>
-								</Card.Content>
-							</Card>
-							<Card>
-								<Card.Content>
-									<Image floated='right' src={`${process.env.GATSBY_ASSETS_URL}atlas/icon_${voyageConfig.skills.secondary_skill}.png`} style={{ height: '2em' }} />
-									<Card.Header>{CONFIG.SKILLS[voyageConfig.skills.secondary_skill]}</Card.Header>
-									<p>secondary</p>
-								</Card.Content>
-							</Card>
-							<Card>
-								<Card.Content>
-									<Card.Header>{allTraits.ship_trait_names[voyageConfig.ship_trait] ?? voyageConfig.ship_trait}</Card.Header>
-									<p>ship trait</p>
-								</Card.Content>
-							</Card>
-						</Card.Group>
-					</Grid.Column>
-					<Grid.Column width={2} textAlign='right'>
-						<VoyageEditConfigModal voyageConfig={voyageConfig} updateConfig={updateConfig} />
-					</Grid.Column>
-				</Grid>
-			}
-
-			{voyageState !== 'input' && (<VoyageExisting voyageConfig={voyageConfig} useCalc={() => setVoyageState('input')} roster={myCrew} />)}
-			{voyageState === 'input' && (<VoyageInput voyageConfig={voyageConfig} myCrew={myCrew} useInVoyage={() => setVoyageState(voyageConfig.state)} />)}
+			<Grid columns={2} stackable>
+				<Grid.Column width={14}>
+					<Card.Group>
+						<Card>
+							<Card.Content>
+								<Image floated='right' src={`${process.env.GATSBY_ASSETS_URL}atlas/icon_${voyageConfig.skills.primary_skill}.png`} style={{ height: '2em' }} />
+								<Card.Header>{CONFIG.SKILLS[voyageConfig.skills.primary_skill]}</Card.Header>
+								<p>primary</p>
+							</Card.Content>
+						</Card>
+						<Card>
+							<Card.Content>
+								<Image floated='right' src={`${process.env.GATSBY_ASSETS_URL}atlas/icon_${voyageConfig.skills.secondary_skill}.png`} style={{ height: '2em' }} />
+								<Card.Header>{CONFIG.SKILLS[voyageConfig.skills.secondary_skill]}</Card.Header>
+								<p>secondary</p>
+							</Card.Content>
+						</Card>
+						<Card>
+							<Card.Content>
+								<Card.Header>{allTraits.ship_trait_names[voyageConfig.ship_trait] ?? voyageConfig.ship_trait}</Card.Header>
+								<p>ship trait</p>
+							</Card.Content>
+						</Card>
+					</Card.Group>
+				</Grid.Column>
+				<Grid.Column width={2} textAlign='right'>
+					{voyageState === 'input' &&
+						<ConfigEditor voyageConfig={voyageConfig} updateConfig={updateConfig} />
+					}
+				</Grid.Column>
+			</Grid>
+			<div style={{ marginTop: '1em' }}>
+				{voyageState !== 'input' && (<VoyageExisting voyageConfig={voyageConfig} useCalc={() => setVoyageState('input')} roster={myCrew} />)}
+				{voyageState === 'input' && (<VoyageInput voyageConfig={voyageConfig} myCrew={myCrew} useInVoyage={() => setVoyageState(voyageConfig.state)} />)}
+			</div>
 		</React.Fragment>
 	);
 
@@ -200,201 +198,6 @@ const VoyageMain = (props: VoyageMainProps) => {
 	}
 };
 
-type VoyageEditConfigModalProps = {
-	voyageConfig: any;
-	updateConfig: (newVoyageConfig: any) => void;
-};
-
-const VoyageEditConfigModal = (props: VoyageEditConfigModalProps) => {
-	const { updateConfig } = props;
-
-	const [voyageConfig, setVoyageConfig] = React.useState(props.voyageConfig);
-
-	const [modalIsOpen, setModalIsOpen] = React.useState(false);
-	const [updateOnClose, setUpdateOnClose] = React.useState(false);
-	const [options, setOptions] = React.useState(undefined);
-
-	React.useEffect(() => {
-		if (!modalIsOpen && updateOnClose) {
-			updateConfig(voyageConfig);
-			setUpdateOnClose(false);
-		}
-
-	}, [modalIsOpen]);
-
-	const defaultSlots = [
-		{ symbol: 'captain_slot', name: 'First Officer', skill: 'command_skill', trait: '' },
-		{ symbol: 'first_officer', name: 'Helm Officer', skill: 'command_skill', trait: '' },
-		{ symbol: 'chief_communications_officer', name: 'Communications Officer', skill: 'diplomacy_skill', trait: '' },
-		{ symbol: 'communications_officer', name: 'Diplomat', skill: 'diplomacy_skill', trait: '' },
-		{ symbol: 'chief_security_officer', name: 'Chief Security Officer', skill: 'security_skill', trait: '' },
-		{ symbol: 'security_officer', name: 'Tactical Officer', skill: 'security_skill', trait: '' },
-		{ symbol: 'chief_engineering_officer', name: 'Chief Engineer', skill: 'engineering_skill', trait: '' },
-		{ symbol: 'engineering_officer', name: 'Engineer', skill: 'engineering_skill', trait: '' },
-		{ symbol: 'chief_science_officer', name: 'Chief Science Officer', skill: 'science_skill', trait: '' },
-		{ symbol: 'science_officer', name: 'Deputy Science Officer', skill: 'science_skill', trait: '' },
-		{ symbol: 'chief_medical_officer', name: 'Chief Medical Officer', skill: 'medicine_skill', trait: '' },
-		{ symbol: 'medical_officer', name: 'Ship\'s Counselor', skill: 'medicine_skill', trait: '' }
-	];
-	const crewSlots = voyageConfig.crew_slots ?? defaultSlots;
-	crewSlots.sort((s1, s2) => CONFIG.VOYAGE_CREW_SLOTS.indexOf(s1.symbol) - CONFIG.VOYAGE_CREW_SLOTS.indexOf(s2.symbol));
-
-	return (
-		<Modal
-			open={modalIsOpen}
-			onClose={() => setModalIsOpen(false)}
-			onOpen={() => setModalIsOpen(true)}
-			trigger={<Button size='small'><Icon name='edit' />Edit</Button>}
-		>
-			<Modal.Header>Edit Voyage</Modal.Header>
-			<Modal.Content scrolling>
-				{renderContent()}
-			</Modal.Content>
-			<Modal.Actions>
-				<Button positive onClick={() => setModalIsOpen(false)}>
-					Close
-				</Button>
-			</Modal.Actions>
-		</Modal>
-	);
-
-	function renderContent(): JSX.Element {
-		if (!modalIsOpen) return (<></>);
-
-		if (!options) {
-			// Renders a lot faster by using known voyage traits rather than calculate list from all possible traits
-			const knownShipTraits = ['andorian','battle_cruiser','borg','breen','cardassian','cloaking_device',
-				'dominion','emp','explorer','federation','ferengi','fighter','freighter','historic','hologram',
-				'klingon','malon','maquis','orion_syndicate','pioneer','reman','romulan','ruthless',
-				'scout','sikarian','spore_drive','terran','tholian','transwarp','vulcan','warship','war_veteran','xindi'];
-			const knownCrewTraits = ['android','astrophysicist','bajoran','borg','brutal',
-				'cardassian','civilian','communicator','costumed','crafty','cultural_figure','cyberneticist',
-				'desperate','diplomat','doctor','duelist','exobiology','explorer','federation','ferengi',
-				'gambler','hero','hologram','human','hunter','innovator','inspiring','jury_rigger','klingon',
-				'marksman','maverick','pilot','prodigy','resourceful','romantic','romulan',
-				'saboteur','scoundrel','starfleet','survivalist','tactician','telepath','undercover_operative',
-				'veteran','villain','vulcan'];
-
-			const skillsList = [];
-			for (let skill in CONFIG.SKILLS) {
-				skillsList.push({
-					key: skill,
-					value: skill,
-					text: CONFIG.SKILLS[skill]
-				});
-			}
-
-			const shipTraitsList = knownShipTraits.map(trait => {
-				return {
-					key: trait,
-					value: trait,
-					text: allTraits.ship_trait_names[trait] ?? trait
-				};
-			});
-			shipTraitsList.sort((a, b) => a.text.localeCompare(b.text));
-
-			const crewTraitsList = knownCrewTraits.map(trait => {
-				return {
-					key: trait,
-					value: trait,
-					text: allTraits.trait_names[trait]
-				};
-			});
-			crewTraitsList.sort((a, b) => a.text.localeCompare(b.text));
-
-			setOptions({ skills: skillsList, ships: shipTraitsList, traits: crewTraitsList });
-			return (<></>);
-		}
-
-		return (
-			<React.Fragment>
-				<Message>Editing this voyage will reset all existing recommendations and estimates.</Message>
-				<Form>
-					<Form.Group>
-						<Form.Select
-							label='Primary skill'
-							options={options.skills}
-							value={voyageConfig.skills.primary_skill ?? 'command_skill'}
-							onChange={(e, { value }) => setSkill('primary_skill', value)}
-							placeholder='Primary'
-						/>
-						<Form.Select
-							label='Secondary skill'
-							options={options.skills}
-							value={voyageConfig.skills.secondary_skill ?? 'science_skill'}
-							onChange={(e, { value }) => setSkill('secondary_skill', value)}
-							placeholder='Secondary'
-						/>
-						<Form.Select
-							search clearable
-							label='Ship trait'
-							options={options.ships}
-							value={voyageConfig.ship_trait}
-							onChange={(e, { value }) => setShipTrait(value)}
-							placeholder='Ship trait'
-						/>
-					</Form.Group>
-				</Form>
-				<Table compact striped>
-					<Table.Header>
-						<Table.Row>
-							<Table.HeaderCell textAlign='center'>Skill</Table.HeaderCell>
-							<Table.HeaderCell>Seat</Table.HeaderCell>
-							<Table.HeaderCell>Trait</Table.HeaderCell>
-						</Table.Row>
-					</Table.Header>
-					<Table.Body>
-						{crewSlots.map((seat, idx) => (
-							<Table.Row key={seat.symbol}>
-								{ idx % 2 == 0 ?
-									(
-										<Table.Cell rowSpan='2' textAlign='center'>
-											<img alt="{seat.skill}" src={`${process.env.GATSBY_ASSETS_URL}atlas/icon_${seat.skill}.png`} style={{ height: '2em' }} />
-										</Table.Cell>
-									)
-									: (<></>)
-								}
-								<Table.Cell>{seat.name}</Table.Cell>
-								<Table.Cell>
-									<Dropdown search selection clearable
-										options={options.traits}
-										value={seat.trait}
-										onChange={(e, { value }) => setSeatTrait(seat.symbol, value)}
-										placeholder='Trait'
-									/>
-								</Table.Cell>
-							</Table.Row>
-						))}
-					</Table.Body>
-				</Table>
-			</React.Fragment>
-		);
-	}
-
-	function setSkill(prime: string, value: string): void {
-		// Flip skill values if changing to value that's currently set as the other prime
-		if (prime == 'primary_skill' && value == voyageConfig.skills.secondary_skill)
-			voyageConfig.skills.secondary_skill = voyageConfig.skills.primary_skill;
-		else if (prime == 'secondary_skill' && value == voyageConfig.skills.primary_skill)
-			voyageConfig.skills.primary_skill = voyageConfig.skills.secondary_skill;
-		voyageConfig.skills[prime] = value;
-		setVoyageConfig({...voyageConfig});
-		setUpdateOnClose(true);
-	}
-
-	function setShipTrait(value: string): void {
-		voyageConfig.ship_trait = value;
-		setVoyageConfig({...voyageConfig});
-		setUpdateOnClose(true);
-	}
-
-	function setSeatTrait(seat: symbol, value: string): void {
-		voyageConfig.crew_slots.find(s => s.symbol === seat).trait = value;
-		setVoyageConfig({...voyageConfig});
-		setUpdateOnClose(true);
-	}
-};
-
 type VoyageExistingProps = {
 	voyageConfig: any;
 	useCalc: () => void;
@@ -406,7 +209,7 @@ const VoyageExisting = (props: VoyageExistingProps) => {
 	const { voyageConfig, useCalc, roster } = props;
 
 	return (
-		<div style={{ marginTop: '1em' }}>
+		<React.Fragment>
 			<VoyageStats
 				voyageData={voyageConfig}
 				ships={allShips}
@@ -414,109 +217,15 @@ const VoyageExisting = (props: VoyageExistingProps) => {
 				playerItems={playerData.player.character.items}
 				roster={roster}
 			/>
-			<div style={{ marginTop: '1em' }}>
-				<Button onClick={() => useCalc()}>Return to crew calculator</Button>
-			</div>
 			{voyageConfig.state !== 'pending' && <CIVASMessage voyageConfig={voyageConfig} />}
-		</div>
-	)
-};
-
-type CIVASMessageProps = {
-	voyageConfig: any;
-};
-
-const CIVASMessage = (props: CIVASMessageProps) => {
-	const CIVASLink = 'https://docs.google.com/spreadsheets/d/1utIuwIgIRO7mwYMSP3P9eWEygf1YT9k24A8BZ_l4pOw/edit?usp=sharing';
-	const CIVASVer = '2.1';
-
-	enum ExportState {
-		None,
-		InProgress,
-		Ready
-	};
-
-	const { voyageConfig } = props;
-	const [exportState, setExportState] = React.useState(ExportState.None);
-	const [estimate, setEstimate] = React.useState(undefined);
-
-	React.useEffect(() => {
-		if (estimate) copyToClipboard();
-	}, [estimate]);
-
-	if (!navigator.clipboard) return (<></>);
-
-	return (
-		<React.Fragment>
-			<Message style={{ marginTop: '2em' }}>
-				<Message.Content>
-					<Message.Header>Captain Idol's Voyage Analysis Sheet</Message.Header>
-					<p>Pro tip: use <b><a href={CIVASLink} target='_blank'>Captain Idol's Voyage Analysis Sheet</a></b> to help you keep track of your voyagers, runtimes, and estimates. Click the button below to copy the details of your current voyage to the clipboard so that the relevant data can be pasted directly into your CIVAS Google Sheet (currently v{CIVASVer}).</p>
-					<Popup
-						content={exportState === ExportState.Ready ? 'Copied!' : 'Please wait...'}
-						on='click'
-						position='right center'
-						size='tiny'
-						trigger={
-							<Button icon='clipboard check' content='Copy details to clipboard' onClick={() => exportData()} />
-						}
-					/>
-				</Message.Content>
-			</Message>
+			<Button content='Return to crew calculator'
+				icon='exchange'
+				size='large'
+				onClick={()=> useCalc()}
+				style={{ marginTop: '2em' }}
+			/>
 		</React.Fragment>
-	);
-
-	function exportData(): void {
-		if (exportState === ExportState.InProgress) return;
-		if (estimate) {
-			copyToClipboard();
-			return;
-		}
-
-		setExportState(ExportState.InProgress);
-		const config = {
-			startAm: voyageConfig.max_hp,
-			ps: voyageConfig.skill_aggregates[voyageConfig.skills['primary_skill']],
-			ss: voyageConfig.skill_aggregates[voyageConfig.skills['secondary_skill']],
-			others: Object.values(voyageConfig.skill_aggregates).filter(s => !Object.values(voyageConfig.skills).includes(s.skill))
-		};
-		const ChewableConfig = {
-			config,
-			worker: 'chewable'
-		};
-		const worker = new UnifiedWorker();
-		worker.addEventListener('message', message => {
-			if (!message.data.inProgress) {
-				setExportState(ExportState.Ready);
-				setEstimate(message.data.result);
-			}
-		});
-		worker.postMessage(ChewableConfig);
-	}
-
-	function copyToClipboard(): void {
-		const hoursToTime = hours => {
-			let wholeHours = Math.floor(hours);
-			return `${wholeHours}:${Math.floor((hours-wholeHours)*60).toString().padStart(2, '0')}`
-		};
-
-		const skillToShort = skillName => CONFIG.SKILLS_SHORT.find(skill => skill.name === skillName).short;
-
-		let values = [
-			skillToShort(voyageConfig.skills['primary_skill'])+'/'+skillToShort(voyageConfig.skills['secondary_skill']),
-			new Date(voyageConfig.created_at).toISOString().split('T')[0],
-			hoursToTime(estimate.refills[0].result)
-		];
-		values.push(voyageConfig.state === 'recalled' ? hoursToTime(voyageConfig.log_index/180) : '');
-		values.push(voyageConfig.state === 'recalled' ? voyageConfig.hp : '');
-		values = values.concat(voyageConfig
-			.crew_slots
-			.sort((s1, s2) => CONFIG.VOYAGE_CREW_SLOTS.indexOf(s1.symbol) - CONFIG.VOYAGE_CREW_SLOTS.indexOf(s2.symbol))
-			.map(s => s.crew.name)
-		);
-
-		navigator.clipboard.writeText(values.join('\t'));
-	}
+	)
 };
 
 type VoyageInputProps = {
@@ -526,529 +235,24 @@ type VoyageInputProps = {
 };
 
 const VoyageInput = (props: VoyageInputProps) => {
-	const { allShips, playerData } = React.useContext(AllDataContext);
-	const { voyageConfig, myCrew, useInVoyage } = props;
-
-	const [bestShip, setBestShip] = React.useState(undefined);
-	const [consideredCrew, setConsideredCrew] = React.useState([]);
-	const [calculator, setCalculator] = useStateWithStorage(playerData.player.dbid+'/voyage/calculator', 'iampicard', { rememberForever: true });
-	const [calcOptions, setCalcOptions] = useStateWithStorage(playerData.player.dbid+'/voyage/calcOptions', {}, { rememberForever: true });
-	const [telemetryOptOut, setTelemetryOptOut] = useStateWithStorage('telemetryOptOut', false, { rememberForever: true });
-	const [requests, setRequests] = React.useState([]);
-	const [results, setResults] = React.useState([]);
-
-	React.useEffect(() => {
-		const consideredShips = [];
-		allShips.filter(ship => ship.owned).forEach(ship => {
-			const traited = ship.traits.includes(voyageConfig.ship_trait);
-			let entry = {
-				ship: ship,
-				score: ship.antimatter + (traited ? 150 : 0),
-				traited: traited,
-				bestIndex: Math.min(ship.index.left, ship.index.right)
-			};
-			consideredShips.push(entry);
-		});
-		consideredShips.sort((a, b) => {
-			if (a.score === b.score) return a.bestIndex - b.bestIndex;
-			return b.score - a.score;
-		});
-		setBestShip(consideredShips[0]);
-		setRequests([]);
-		setResults([]);
-	}, [voyageConfig]);
-
-	React.useEffect(() => {
-		return function cleanup() {
-			// Cancel active calculations when leaving page
-			requests.forEach(request => {
-				if (request.calcState == CalculatorState.InProgress)
-					request.abort();
-			});
-		}
-	}, []);
-
-	// Scroll here when calculator started, finished
-	const topAnchor = React.useRef(null);
-
-	const calculators = CALCULATORS.helpers.map(helper => {
-		return { key: helper.id, value: helper.id, text: helper.name };
-	});
-	calculators.push({ key: 'all', value: 'all', text: 'All calculators (slower)' });
-
-	return (
-		<React.Fragment>
-			<div ref={topAnchor} />
-			{renderBestShip()}
-			{renderResults()}
-			{requests.length > 0 && <Header as='h3'>Options</Header>}
-			<Form>
-				<InputCrewOptions myCrew={myCrew} updateConsideredCrew={setConsideredCrew} />
-				<Form.Group inline>
-					<Form.Field
-						control={Select}
-						label='Calculator'
-						options={calculators}
-						value={calculator}
-						onChange={(e, { value }) => setCalculator(value)}
-						placeholder='Select calculator'
-					/>
-					{CALCULATORS.fields.filter(field => field.calculators.includes(calculator) || calculator == 'all').map(field => (
-						<Form.Field
-							key={field.id}
-							control={Select}	/* Only control allowed at the moment */
-							label={field.name}
-							options={field.options}
-							value={calcOptions[field.id] ?? field.default}
-							placeholder={field.description}
-							onChange={(e, { value }) => setCalcOptions(prevOptions =>
-								{
-									const newValue = { [field.id]: value };
-									return {...prevOptions, ...newValue};
-								}
-							)}
-						/>
-					))}
-				</Form.Group>
-
-				<Form.Group>
-					<Form.Button primary onClick={() => startCalculation()}>
-						Calculate best crew selection
-					</Form.Button>
-					{voyageConfig.state &&
-						<Form.Button onClick={()=> useInVoyage()}>
-							Return to current voyage
-						</Form.Button>
-					}
-				</Form.Group>
-			</Form>
-			<Message style={{ marginTop: '2em' }}>
-				<Message.Content>
-					<Message.Header>Privacy Notice</Message.Header>
-					<p>We use anonymous statistics aggregated from voyage calculations to improve DataCore and power our <b><Link to='/hall_of_fame'>Voyage Hall of Fame</Link></b>.</p>
-					<Form>
-						<Form.Field
-							control={Checkbox}
-							label={<label>Permit DataCore to collect anonymous voyage stats</label>}
-							checked={!telemetryOptOut}
-							onChange={(e, { checked }) => setTelemetryOptOut(!checked) }
-						/>
-					</Form>
-				</Message.Content>
-			</Message>
-		</React.Fragment>
-	);
-
-	function renderBestShip(): JSX.Element {
-		if (!bestShip) return (<></>);
-
-		const direction = bestShip.ship.index.right < bestShip.ship.index.left ? 'right' : 'left';
-		const index = bestShip.ship.index[direction] ?? 0;
-
-		return (
-			<Card fluid>
-				<Card.Content>
-					<Image floated='left' src={`${process.env.GATSBY_ASSETS_URL}${bestShip.ship.icon.file.substr(1).replace('/', '_')}.png`} style={{ height: '4em' }} />
-					<Card.Header>{bestShip.ship.name}</Card.Header>
-					<p>best ship{bestShip.traited && (<span style={{ marginLeft: '1em' }}>{` +`}{allTraits.ship_trait_names[voyageConfig.ship_trait]}</span>)}</p>
-					<p style={{ marginTop: '.5em' }}>Tap <Icon name={`arrow ${direction}`} />{index} time{index != 1 ? 's' : ''} on your voyage ship selection screen to select {bestShip.ship.name}.</p>
-				</Card.Content>
-			</Card>
-		);
-	}
-
-	function renderResults(): JSX.Element {
-		if (results.length == 0)
-			return (<></>);
-
-		// In-game voyage crew picker ignores immortalized crew and crew active on shuttles
-		const availableRoster = myCrew.filter(c => c.immortal === 0 && c.active_status !== 2);
-
-		const showPopup = (result) => <Popup basic content={<p>{result.result.postscript}</p>} trigger={<p>{result.name}</p>} />
-		const panes = results.map(result => ({
-			menuItem: { key: result.id, content: result.result ? showPopup(result) : result.name },
-			render: () => (
-				<VoyageResultPane result={result.result}
-					requests={requests} requestId={result.requestId}
-					calcState={result.calcState} abortCalculation={abortCalculation}
-					roster={availableRoster}
-				/>
-			)
-		}));
-
-		return (
-			<React.Fragment>
-				<Header as='h3'>Recommended Lineups</Header>
-				<Tab menu={{ pointing: true }}
-					panes={panes}
-				/>
-			</React.Fragment>
-		);
-	}
-
-	function scrollToAnchor(): void {
-		if (!topAnchor.current) return;
-		topAnchor.current.scrollIntoView({
-			behavior: 'smooth'
-		}, 500);
-	}
-
-	function startCalculation(): void {
-		const helperConfig = {
-			voyageConfig, bestShip, consideredCrew, calcOptions,
-			resultsCallback: handleResults,
-			isMobile
-		};
-		CALCULATORS.helpers.forEach(helper => {
-			if (helper.id == calculator || calculator == 'all') {
-				const request = helper.helper(helperConfig);
-				requests.push(request);
-				results.push({
-					id: request.id,
-					requestId: request.id,
-					name: 'Calculating...',
-					calcState: CalculatorState.InProgress
-				});
-				request.start();
-			}
-		});
-		setRequests([...requests]);
-		setResults([...results]);
-		scrollToAnchor();
-	}
-
-	function handleResults(requestId: string, reqResults: any[], calcState: number): void {
-		reqResults.forEach((reqResult, idx) => {
-			// Update existing pane with results
-			if (idx == 0) {
-				setResults(prevResults => {
-					const result = prevResults.find(r => r.id == requestId);
-					if (calcState == CalculatorState.Done) {
-						result.name = formatTime(reqResult.estimate.refills[0].result);
-						result.calcState = CalculatorState.Done;
-						sendTelemetry(requestId, reqResult);
-					}
-					result.result = reqResult;
-					return [...prevResults];
-				});
-			}
-			// Add new panes if multiple results generated by this request
-			else {
-				setResults(prevResults => [...prevResults, {
-					id: requestId+'-'+idx,
-					requestId,
-					name: formatTime(reqResult.estimate.refills[0].result),
-					calcState: CalculatorState.Done,
-					result: reqResult
-				}]);
-			}
-		});
-		if (calcState == CalculatorState.Done) scrollToAnchor();
-	}
-
-	function abortCalculation(requestId: string): void {
-		const request = requests.find(r => r.id == requestId);
-		if (request) {
-			request.abort();
-			setResults(prevResults => {
-				const result = prevResults.find(prev => prev.id == requestId);
-				if (result.result) {
-					result.name = formatTime(result.result.estimate.refills[0].result);
-					result.calcState = CalculatorState.Done;
-				}
-				else {
-					const index = prevResults.findIndex(prev => prev.id == requestId);
-					prevResults.splice(index, 1);
-				}
-				return [...prevResults];
-			});
-		}
-	}
-
-	function sendTelemetry(requestId: string, result: any): void {
-		if (telemetryOptOut) return;
-		const request = requests.find(r => r.id == requestId);
-		const estimatedDuration = result.estimate.refills[0].result*60*60;
-		try {
-			fetch(`${process.env.GATSBY_DATACORE_URL}api/telemetry`, {
-				method: 'post',
-				headers: {
-					'Content-Type': 'application/json'
-				},
-				body: JSON.stringify({
-					type: 'voyageCalc',
-					data: {
-						voyagers: result.entries.map((entry) => entry.choice.symbol),
-						estimatedDuration,
-						calculator: request ? request.calculator : ''
-					}
-				})
-			});
-		}
-		catch (err) {
-			console.log('An error occurred while sending telemetry', err);
-		}
-	}
-};
-
-type InputCrewOptionsProps = {
-	myCrew: any[];
-	updateConsideredCrew: () => void;
-};
-
-const InputCrewOptions = (props: InputCrewOptionsProps) => {
-	const { myCrew, updateConsideredCrew } = props;
-
-	const [considerActive, setConsiderActive] = React.useState(false);
-	const [considerFrozen, setConsiderFrozen] = React.useState(false);
-	const [excludedCrew, setExcludedCrew] = React.useState([]);
-
-	React.useEffect(() => {
-		const consideredCrew = myCrew.filter(crewman => {
-			if (!considerActive && crewman.active_status == 2)
-				return false;
-
-			if (!considerFrozen && crewman.immortal > 0)
-				return false;
-
-			if (excludedCrew && excludedCrew.includes(crewman.id))
-				return false;
-
-			return true;
-		});
-		updateConsideredCrew(consideredCrew);
-	}, [considerActive, considerFrozen, excludedCrew]);
-
-	const activeCount = myCrew.filter(crew => crew.active_status == 2).length;
-
-	return (
-		<React.Fragment>
-			<Form.Group grouped>
-				{activeCount > 0 && (
-					<Form.Field
-						control={Checkbox}
-						label='Consider crew on active shuttles'
-						checked={considerActive}
-						onChange={(e, { checked }) => setConsiderActive(checked)}
-					/>
-				)}
-				<Form.Field
-					control={Checkbox}
-					label='Consider frozen (vaulted) crew'
-					checked={considerFrozen}
-					onChange={(e, { checked }) => setConsiderFrozen(checked)}
-				/>
-			</Form.Group>
-			<InputCrewExcluder myCrew={myCrew} excludedCrew={excludedCrew} updateExclusions={setExcludedCrew} showFrozen={considerFrozen} />
-		</React.Fragment>
-	);
-};
-
-type InputCrewExcluderProps = {
-	myCrew: any[];
-	excludedCrew: undefined | number[];
-	updateExclusions: (crewIds: number[]) => void;
-	showFrozen: boolean;
-}
-
-const InputCrewExcluder = (props: InputCrewExcluderProps) => {
-	const { allCrew } = React.useContext(AllDataContext);
-	const { updateExclusions } = props;
-
-	const [eventData, setEventData] = useStateWithStorage('tools/eventData', undefined);
-	const [activeEvent, setActiveEvent] = React.useState(undefined);
-	const [options, setOptions] = React.useState(undefined);
-
-	React.useEffect(() => {
-		if (props.excludedCrew)
-			if (options && !options.initialized) populatePlaceholders();
-	}, [props.excludedCrew]);
-
-	React.useEffect(() => {
-		if (!options) return;
-		if (!options.initialized)
-			populatePlaceholders();
-		else
-			populateOptions();
-	}, [props.showFrozen]);
-
-	React.useEffect(() => {
-		if (activeEvent && activeEvent.seconds_to_end > 0 && activeEvent.seconds_to_start < 86400) {
-			if (activeEvent.content_types.includes('shuttles') || activeEvent.content_types.includes('gather')) {
-				const crewIds = props.myCrew.filter(c => activeEvent.bonus.includes(c.symbol)).map(c => c.id);
-				updateExclusions([...crewIds]);
-			}
-		}
-	}, [activeEvent]);
-
-	if (!activeEvent) {
-		identifyActiveEvent();
-		return (<></>);
-	}
-
-	if (!options) {
-		populatePlaceholders();
-		return (<></>);
-	}
-
-	const label = (
-		<React.Fragment>
-			<label>Crew to exclude from voyage</label>
-			{activeEvent && activeEvent.bonus.length > 0 &&
-				(<div style={{ margin: '-.5em 0 .5em' }}>Preselected crew give bonuses for the event <b>{activeEvent.name}</b></div>)
-			}
-		</React.Fragment>
-	);
-
-	return (
-		<React.Fragment>
-			<Form.Field
-				label={label}
-				placeholder='Search for crew to exclude'
-				control={Dropdown}
-				clearable
-				fluid
-				multiple
-				search
-				selection
-				options={options.list}
-				value={props.excludedCrew}
-				onFocus={() => { if (!options.initialized) populateOptions(); }}
-				onChange={(e, { value }) => updateExclusions(value) }
-			/>
-		</React.Fragment>
-	);
-
-	function identifyActiveEvent(): void {
-		// Get event data from recently uploaded playerData
-		if (eventData && eventData.length > 0) {
-			const currentEvent = getEventData(eventData.sort((a, b) => (a.seconds_to_start - b.seconds_to_start))[0], allCrew);
-			setActiveEvent({...currentEvent});
-		}
-		// Otherwise guess event from autosynced events
-		/* 	Uncomment when voyagecalc can run without playerData
-		else {
-			guessCurrentEvent().then(currentEvent => {
-				setActiveEvent({...currentEvent});
-			});
-		}
-		*/
-	}
-
-	function populatePlaceholders(): void {
-		const options = { initialized: false, list: [] };
-		if (props.excludedCrew.length > 0) {
-			let crewList = [...props.myCrew];
-			if (!props.showFrozen) crewList = crewList.filter(c => c.immortal == 0);
-			options.list = crewList.filter(c => props.excludedCrew.includes(c.id)).map(c => {
-				return { key: c.id, value: c.id, text: c.name, image: { avatar: true, src: `${process.env.GATSBY_ASSETS_URL}${c.imageUrlPortrait}` }};
-			});
-		}
-		else {
-			options.list = [{ key: 0, value: 0, text: 'Loading...' }];
-		}
-		setOptions({...options});
-	}
-
-	function populateOptions(): void {
-		let crewList = [...props.myCrew];
-		if (!props.showFrozen) crewList = crewList.filter(c => c.immortal == 0);
-		options.list = crewList.sort((a, b) => a.name.localeCompare(b.name)).map(c => {
-			return { key: c.id, value: c.id, text: c.name, image: { avatar: true, src: `${process.env.GATSBY_ASSETS_URL}${c.imageUrlPortrait}` }};
-		});
-		options.initialized = true;
-		setOptions({...options});
-	}
-};
-
-type VoyageResultPaneProps = {
-	result: any;
-	requests: any[];
-	requestId: string;
-	calcState: number;
-	abortCalculation: (requestId: string) => void;
-	roster: any[];
-};
-
-const VoyageResultPane = (props: VoyageResultPaneProps) => {
-	const { result, requests, requestId, calcState, abortCalculation, roster } = props;
-
-	const request = requests.find(r => r.id == requestId);
-	if (!request) return (<></>);
-
-	if (!result) {
-		return (
-			<Tab.Pane>
-				<div style={{ textAlign: 'center' }}>
-					<Image centered src='/media/voyage-wait-icon.gif' />
-					<Button onClick={() => abortCalculation(request.id)}>Abort</Button>
-				</div>
-			</Tab.Pane>
-		);
-	}
-
-	// resultToVoyageData
-	let data = {...request.voyageConfig};
-	if (result.entries) {
-		result.entries.forEach((entry, idx) => {
-			let acrew = request.consideredCrew.find(c => c.id === entry.choice.id);
-			data.crew_slots[entry.slotId].crew = acrew;
-		});
-	}
-	data.skill_aggregates = result.aggregates;
-	data.max_hp = result.startAM;
-	data.state = 'pending';
-
-	const renderCalculatorMessage = () => {
-		if (calcState != CalculatorState.Done) {
-			return (
-				<>
-					<Image inline size='mini' src='/media/voyage-wait-icon.gif' />
-					Calculation in progress. Please wait...{` `}
-					<Button size='mini' onClick={() => abortCalculation(request.id)}>Abort</Button>
-				</>
-			);
-		}
-		let inputs = Object.entries(request.calcOptions).map(entry => entry[0]+': '+entry[1]);
-		inputs.unshift('considered crew: '+request.consideredCrew.length);
-		return (
-			<>
-				Calculated by <b>{request.calcName}</b> calculator ({inputs.join(', ')}){` `}
-				in {((request.perf.end-request.perf.start)/1000).toFixed(2)} seconds!
-				{result.postscript && (<div>{result.postscript}</div>)}
-			</>
-		);
+	const { allCrew, allShips, playerData } = React.useContext(AllDataContext);
+	const { voyageConfig } = props;
+	const allData = {
+		allCrew, allShips, playerData
 	};
-
 	return (
 		<React.Fragment>
-			{calcState == CalculatorState.Done && (
-				<Message attached>
-					Estimate: <b>{formatTime(result.estimate.refills[0].result)}</b>{` `}
-					(expected range: {formatTime(result.estimate.refills[0].saferResult)} to{` `}
-						{formatTime(result.estimate.refills[0].moonshotResult)})
-				</Message>
-			)}
-			<Tab.Pane>
-				<VoyageStats
-					voyageData={data}
-					estimate={result.estimate}
-					ships={[request.bestShip]}
-					roster={roster}
-					showPanels={['crew']}
+			<Recommender voyageConfig={props.voyageConfig} myCrew={props.myCrew} allData={allData} />
+			{voyageConfig.state &&
+				<Button content='Return to current voyage'
+					icon='exchange'
+					size='large'
+					onClick={()=> props.useInVoyage()}
+					style={{ marginTop: '2em' }}
 				/>
-				<div style={{ marginTop: '1em' }}>
-					{renderCalculatorMessage()}
-				</div>
-			</Tab.Pane>
+			}
 		</React.Fragment>
 	);
-};
-
-const formatTime: string = (time: number) => {
-	let hours = Math.floor(time);
-	let minutes = Math.floor((time-hours)*60);
-	return hours+"h " +minutes+"m";
 };
 
 export default VoyageCalculator;
